My_web/requirements.txt
flask
numpy
Pillow
pandas
ipython
matplotlib
seaborn
opencv-python-headless
torchvision
torchaudio
torch